🔨 Refr: System instances now init plants within stars
When Star instances init they now create planet instance variables. This means that when creating a system instance and stars are generated, they generate their own planets.

Planet instance generation has been removed from System instances
Star instances generate planet instances

All object referencing throughout the website has been upated to remove old system.planets references and now alike to system.star.planets
