+++ templates/new_planets.html
@@ -8,11 +8,11 @@
         <thead>
             <tr>
                 <th></th>
-                <th>Planet Name</th>
-                <th>Host</th>
-                <th>System</th>
-                <th>Circumbinary?</th>
-                <th>Public Discovery Date</th>
+                <th>🪐Planet Name🪐</th>
+                <th>☀️Host☀️</th>
+                <th>🌌System🌌</th>
+                <th>💫Circumbinary?💫</th>
+                <th>📡Public Discovery Date📡</th>
             </tr>
         </thead>
         <tbody>
